apps/chatger/web/pyproject.toml
[tool.poetry]
name = "chatter-judge"
version = "0.1.0"
description = ""
authors = ["Chatter Judge <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.0"
gradio = "3.48.0"
uvicorn = "0.29.0"
itsdangerous = "2.1.2"
sqlalchemy = {extras = ["asyncio"], version = "2.0.28"}
asyncpg = "0.29.0"
google-generativeai = "0.4.1"
python-multipart = "0.0.9"
httpx = "0.27.0"
google-auth-oauthlib = "^1.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "1.0.1"


[tool.poetry.group.dev.dependencies]
mkdocs = "1.5.3"
ruff = "^0.3.5"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
